Fixed warnings found with -Wshorten-64-to-32 flag. 	Change on 2014/06/18 by tball <tball@google.com> ------------- Created by MOE: http://code.google.com/p/moe-java MOE_MIGRATED_REVID=69469620

+++ jre_emul/Classes/DebugUtils.m
@@ -31,7 +31,7 @@ static void LogStack(NSUInteger nFrames) {
   }
 
   void *callStack[nFrames + FRAMES_TO_IGNORE];
-  NSUInteger callFrames = backtrace(callStack, nFrames + FRAMES_TO_IGNORE);
+  NSUInteger callFrames = backtrace(callStack, (int) nFrames + FRAMES_TO_IGNORE);
 
   for (NSUInteger i = FRAMES_TO_IGNORE; i <= callFrames; i++) {
     void *shortStack[1];
